requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==1.10.13
python-dotenv==1.0.0
aiosqlite==0.19.0
motor==3.3.1
pymongo==4.6.0
beanie==1.23.6 